+++ common/src/test/kotlin/combo/math/DataSamplesTest.kt
@@ -14,20 +14,20 @@ class DataSamplesTest {
     fun reservoirSampleFull() {
         val s = ReservoirSample(10, Random)
         for (i in 1..10)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         val collect = s.toArray()
         for (i in 0 until 10)
-            assertEquals(i.toDouble() + 1.0, collect[i])
+            assertEquals(i.toFloat() + 1.0f, collect[i])
     }
 
     @Test
     fun reservoirSampleUnderSize() {
         val s = ReservoirSample(100, Random)
         for (i in 1..10)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         val collect = s.toArray()
         for (i in 0 until 10)
-            assertEquals(i.toDouble() + 1.0, collect[i])
+            assertEquals(i.toFloat() + 1.0f, collect[i])
         assertEquals(10, collect.size)
     }
 
@@ -35,7 +35,7 @@ class DataSamplesTest {
     fun reservoirSampleOverSize() {
         val s = ReservoirSample(20, Random)
         for (i in 1..100)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         val collect = s.toArray()
         assertEquals(20, collect.size)
     }
@@ -49,32 +49,32 @@ class DataSamplesTest {
     fun fullSample() {
         val s = FullSample()
         for (i in 1..100)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         val collect = s.toArray()
         assertEquals(100, collect.size)
         for (i in 1..100)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
     }
 
     @Test
     fun bucketsSampleSizeOne() {
         val s = BucketsSample(1)
         for (i in 1..100)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         val data = s.toArray()
         for (i in 1..100)
-            assertEquals(i.toDouble(), data[i - 1])
+            assertEquals(i.toFloat(), data[i - 1])
     }
 
     @Test
     fun bucketsBig() {
         val s = BucketsSample(10)
         for (i in 1..200)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         val data = s.toArray()
         assertEquals(20, data.size)
         for (i in 0 until 20)
-            assertEquals(5.5 + i * 10.0, data[i])
+            assertEquals(5.5f + i * 10.0f, data[i])
     }
 
     @Test
@@ -86,22 +86,22 @@ class DataSamplesTest {
     fun windowSampleFullSize() {
         val s = WindowSample(20)
         for (i in 1..20)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         val data = s.toArray()
         assertEquals(20, data.size)
         for (i in 1..20)
-            assertEquals(i.toDouble(), data[i - 1])
+            assertEquals(i.toFloat(), data[i - 1])
     }
 
     @Test
     fun windowSampleReUse() {
         val s = WindowSample(20)
         for (i in 1..94)
-            s.accept(i.toDouble())
+            s.accept(i.toFloat())
         s.toArray()
-        s.accept(1.0)
+        s.accept(1.0f)
         val data = s.toArray()
-        assertEquals(1.0, data[data.size - 1])
+        assertEquals(1.0f, data[data.size - 1])
     }
 
     @Test
